refactor(agoric-cli): push follow details down
